rewrite grid spliytter resiz I
